Clean up node-vibrant
