+++ src/panels/lovelace/cards/hui-media-control-card.ts
@@ -16,14 +16,13 @@ import {
 } from "lit-element";
 import { classMap } from "lit-html/directives/class-map";
 import { styleMap } from "lit-html/directives/style-map";
-import Vibrant from "node-vibrant";
-import { Swatch } from "node-vibrant/lib/color";
 import { applyThemesOnElement } from "../../../common/dom/apply_themes_on_element";
 import { fireEvent } from "../../../common/dom/fire_event";
 import { computeStateName } from "../../../common/entity/compute_state_name";
 import { stateIcon } from "../../../common/entity/state_icon";
 import { supportsFeature } from "../../../common/entity/supports-feature";
 import { debounce } from "../../../common/util/debounce";
+import { Vibrant } from "../../../common/image/vibrant";
 import "../../../components/ha-card";
 import "../../../components/ha-icon";
 import "../../../components/ha-icon-button";
@@ -33,7 +32,6 @@ import { UNAVAILABLE_STATES } from "../../../data/entity";
 import {
   computeMediaDescription,
   computeMediaControls,
-  CONTRAST_RATIO,
   getCurrentProgress,
   MediaPickedEvent,
   SUPPORT_BROWSE_MEDIA,
@@ -41,7 +39,6 @@ import {
   SUPPORT_TURN_ON,
 } from "../../../data/media-player";
 import type { HomeAssistant, MediaEntity } from "../../../types";
-import { contrast } from "../common/color/contrast";
 import { findEntities } from "../common/find-entites";
 import { hasConfigOrEntityChanged } from "../common/has-changed";
 import { installResizeObserver } from "../common/install-resize-observer";
@@ -49,114 +46,7 @@ import "../components/hui-marquee";
 import { createEntityNotFoundWarning } from "../components/hui-warning";
 import type { LovelaceCard, LovelaceCardEditor } from "../types";
 import { MediaControlCardConfig } from "./types";
-
-function getContrastRatio(
-  rgb1: [number, number, number],
-  rgb2: [number, number, number]
-): number {
-  return Math.round((contrast(rgb1, rgb2) + Number.EPSILON) * 100) / 100;
-}
-
-// How much the total diff between 2 RGB colors can be
-// to be considered similar.
-const COLOR_SIMILARITY_THRESHOLD = 150;
-
-// For debug purposes, is being tree shaken.
-const DEBUG_COLOR = __DEV__ && false;
-
-const logColor = (
-  color: Swatch,
-  label = `${color.getHex()} - ${color.getPopulation()}`
-) =>
-  // eslint-disable-next-line no-console
-  console.log(
-    `%c${label}`,
-    `color: ${color.getBodyTextColor()}; background-color: ${color.getHex()}`
-  );
-
-const customGenerator = (colors: Swatch[]) => {
-  colors.sort((colorA, colorB) => colorB.population - colorA.population);
-
-  const backgroundColor = colors[0];
-  let foregroundColor: string | undefined;
-
-  const contrastRatios = new Map<Swatch, number>();
-  const approvedContrastRatio = (color: Swatch) => {
-    if (!contrastRatios.has(color)) {
-      contrastRatios.set(
-        color,
-        getContrastRatio(backgroundColor.rgb, color.rgb)
-      );
-    }
-
-    return contrastRatios.get(color)! > CONTRAST_RATIO;
-  };
-
-  // We take each next color and find one that has better contrast.
-  for (let i = 1; i < colors.length && foregroundColor === undefined; i++) {
-    // If this color matches, score, take it.
-    if (approvedContrastRatio(colors[i])) {
-      if (DEBUG_COLOR) {
-        logColor(colors[i], "PICKED");
-      }
-      foregroundColor = colors[i].hex;
-      break;
-    }
-
-    // This color has the wrong contrast ratio, but it is the right color.
-    // Let's find similar colors that might have the right contrast ratio
-
-    const currentColor = colors[i];
-    if (DEBUG_COLOR) {
-      logColor(colors[i], "Finding similar color with better contrast");
-    }
-
-    for (let j = i + 1; j < colors.length; j++) {
-      const compareColor = colors[j];
-
-      // difference. 0 is same, 765 max difference
-      const diffScore =
-        Math.abs(currentColor.rgb[0] - compareColor.rgb[0]) +
-        Math.abs(currentColor.rgb[1] - compareColor.rgb[1]) +
-        Math.abs(currentColor.rgb[2] - compareColor.rgb[2]);
-
-      if (DEBUG_COLOR) {
-        logColor(colors[j], `${colors[j].hex} - ${diffScore}`);
-      }
-
-      if (diffScore > COLOR_SIMILARITY_THRESHOLD) {
-        continue;
-      }
-
-      if (approvedContrastRatio(compareColor)) {
-        if (DEBUG_COLOR) {
-          logColor(compareColor, "PICKED");
-        }
-        foregroundColor = compareColor.hex;
-        break;
-      }
-    }
-  }
-
-  if (foregroundColor === undefined) {
-    foregroundColor = backgroundColor.bodyTextColor;
-  }
-
-  if (DEBUG_COLOR) {
-    // eslint-disable-next-line no-console
-    console.log();
-    // eslint-disable-next-line no-console
-    console.log(
-      "%cPicked colors",
-      `color: ${foregroundColor}; background-color: ${backgroundColor.hex}; font-weight: bold; padding: 16px;`
-    );
-    colors.forEach((color) => logColor(color));
-    // eslint-disable-next-line no-console
-    console.log();
-  }
-
-  return [foregroundColor, backgroundColor.hex];
-};
+import { extractColors } from "../../../common/image/extract_color";
 
 @customElement("hui-media-control-card")
 export class HuiMediaControlCard extends LitElement implements LovelaceCard {
@@ -636,26 +526,21 @@ export class HuiMediaControlCard extends LitElement implements LovelaceCard {
     });
   }
 
-  private _setColors(): void {
+  private async _setColors(): Promise<void> {
     if (!this._image) {
       return;
     }
 
-    new Vibrant(this._image, {
-      colorCount: 16,
-      generator: customGenerator,
-    })
-      .getPalette()
-      .then(([foreground, background]: [string, string]) => {
-        this._backgroundColor = background;
-        this._foregroundColor = foreground;
-      })
-      .catch((err: any) => {
-        // eslint-disable-next-line no-console
-        console.error("Error getting Image Colors", err);
-        this._foregroundColor = undefined;
-        this._backgroundColor = undefined;
-      });
+    try {
+      const { foreground, background } = await extractColors(this._image);
+      this._backgroundColor = background.hex;
+      this._foregroundColor = foreground.hex;
+    } catch (err) {
+      // eslint-disable-next-line no-console
+      console.error("Error getting Image Colors", err);
+      this._foregroundColor = undefined;
+      this._backgroundColor = undefined;
+    }
   }
 
   private _marqueeMouseOver(): void {
